% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1,logical,ANY-method}
\alias{G1,logical,ANY-method}
\title{G1,logical,ANY-method}
\usage{
\S4method{G1}{logical,ANY}(a1, a2, a3, ...)
}
\arguments{
\item{a1}{object of class:\code{logical}, no manual documentation}

\item{a2}{no manual documentation}

\item{a3}{no manual documentation}

\item{...}{no manual documentation}
}
\description{
G1,logical,ANY-method
}
