% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1,numeric,ANY-method}
\alias{G1,numeric,ANY-method}
\title{G1,numeric,ANY-method}
\usage{
\S4method{G1}{numeric,ANY}(a1, a2, a3)
}
\arguments{
\item{a1}{object of class:\code{numeric}, no manual documentation}

\item{a2}{no manual documentation}

\item{a3}{no manual documentation}
}
\description{
G1,numeric,ANY-method
}
