% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1,numeric,character-method}
\alias{G1,numeric,character-method}
\title{G1}
\usage{
\S4method{G1}{numeric,character}(a1, a2, a3)
}
\arguments{
\item{a1}{object of class:\code{numeric}, The only parameter described}

\item{a2}{object of class:\code{character}, no manual documentation}

\item{a3}{no manual documentation}
}
\description{
G1
}
